Refactored benchmark scripts...
... with adjustments due to interface changes
in https://github.com/ttsim-dev/ttsim/pull/42
